refactor: Move all builder level events to useBuilderEvents
